requirements.txt
galaxy.plugin.api==0.69